medical_extractor_project/requirements.txt
streamlit
mysql-connector-python
plotly
langchain~=0.3.20
python-dotenv~=1.0.1
ipykernel
langchain-community~=0.3.19
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters~=0.3.6
langchain-openai~=0.3.8
chromadb==0.6.1
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai~=1.65.4
langchain-groq
duckduckgo-search
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
langchain-core~=0.3.48
uvicorn~=0.34.0
fastapi~=0.115.12
langserve~=0.3.1
dataclasses-json~=0.6.7
streamlit~=1.44.0
pydantic
pydantic-ai
pydantic-ai-slim[openai]
langchain_community
openpyxl~=3.1.0